bug: Remade test project due to assembly error
