+++ MerchStoreTest/Infrastructure/MerchStore.IntegrationTests/appsettings.json
@@ -1,0 +1,18 @@
+{
+    "Logging": {
+      "LogLevel": {
+        "Default": "Information",
+        "Microsoft.AspNetCore": "Warning",
+        "MerchStore.Infrastructure": "Debug",
+        "System.Net.Http.HttpClient": "Warning"
+      }
+    },
+    "ReviewApi": {
+      "BaseUrl": "https://reviewApiv2250506.azurewebsites.net/api/",
+      "ApiKey": "ItHurtWhenIPee",
+      "ApiKeyHeaderName": "x-functions-key",
+      "TimeoutSeconds": 10,
+      "ExceptionsAllowedBeforeBreaking": 3,
+      "CircuitBreakerDurationSeconds": 30
+    }
+  }
