+++ MerchStoreTest/MerchStoreTest.csproj
@@ -1,0 +1,30 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <IsPackable>false</IsPackable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\src\MerchStore.Application\MerchStore.Application.csproj" />
+    <ProjectReference Include="..\src\MerchStore.Domain\MerchStore.Domain.csproj" />
+    <ProjectReference Include="..\src\MerchStore.Infrastructure\MerchStore.Infrastructure.csproj" />
+    <ProjectReference Include="..\src\MerchStore.WebUI\MerchStore.WebUI.csproj" />
+    <ProjectReference Include="..\Infra\ReviewApiClient\ReviewApiClient.csproj" />
+    <ProjectReference Include="..\Infra\ReviewApiFunctionn\ReviewApiFunctions.csproj" />
+
+
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="FluentAssertions" Version="8.2.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="8.1.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="8.1.1" />
+    <PackageReference Include="xunit" Version="2.6.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.8.0" />
+    <PackageReference Include="Moq" Version="4.20.70" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
+    
+  </ItemGroup>
+
+</Project>
