+++ MerchStoreTest/WebUI/Helpers/HoverImageHelperTest.cs
