+++ MerchStoreTest/WebUI/Infrastructure/JsonSnakeCaseNamingPolicyTest.cs
